+++ app/controllers/intouch_cron_jobs_controller.rb
@@ -3,7 +3,7 @@ class IntouchCronJobsController < ApplicationController
 
   layout 'admin'
 
-  before_filter :require_admin
+  before_action :require_admin
 
   accept_api_auth :index