+++ app/controllers/telegram_groups_controller.rb
@@ -3,7 +3,7 @@ class TelegramGroupsController < ApplicationController
 
   layout 'admin'
 
-  before_filter :require_admin
+  before_action :require_admin
 
   def destroy
     TelegramGroupChat.find(params[:id]).destroy