+++ db/migrate/001_create_telegram_users.rb
@@ -1,4 +1,4 @@
-class CreateTelegramUsers < ActiveRecord::Migration
+class CreateTelegramUsers < Rails.version < '5.0' ? ActiveRecord::Migration : ActiveRecord::Migration[4.2]
   def change
     create_table :telegram_users do |t|
       t.integer :tid, index: true