+++ db/migrate/004_create_settings_templates.rb
@@ -1,4 +1,4 @@
-class CreateSettingsTemplates < ActiveRecord::Migration
+class CreateSettingsTemplates < Rails.version < '5.0' ? ActiveRecord::Migration : ActiveRecord::Migration[4.2]
   def change
     create_table :settings_templates do |t|
       t.string :name