+++ db/migrate/006_rename_settings_in_settings_templates.rb
@@ -1,4 +1,4 @@
-class RenameSettingsInSettingsTemplates < ActiveRecord::Migration
+class RenameSettingsInSettingsTemplates < Rails.version < '5.0' ? ActiveRecord::Migration : ActiveRecord::Migration[4.2]
   def change
     rename_column :settings_templates, :settings, :intouch_settings
   end