+++ lib/intouch/patches/projects_helper_patch.rb
@@ -9,7 +9,8 @@ module Intouch
         base.module_eval do
           unloadable
 
-          alias_method_chain :project_settings_tabs, :intouch
+          alias_method :project_settings_tabs_without_intouch, :project_settings_tabs
+          alias_method :project_settings_tabs, :project_settings_tabs_with_intouch
         end
       end
 
@@ -20,10 +21,10 @@ module Intouch
           return tabs unless User.current.allowed_to?(:manage_intouch_settings, @project)
 
           tabs.push(
-            name: 'intouch_settings',
-            action: :manage_intouch_settings,
-            partial: 'projects/settings/intouch/settings',
-            label: 'intouch.label.settings'
+              name: 'intouch_settings',
+              action: :manage_intouch_settings,
+              partial: 'projects/settings/intouch/settings',
+              label: 'intouch.label.settings'
           )
 
           tabs